Renamed button component file used in Book Detail
