+++ src/app/components/BookDetail/BookDetailComponents/BookCover.tsx
@@ -6,8 +6,8 @@ interface BookCoverProps {
 
 const BookCover: React.FC<BookCoverProps> = ({ cover }) => {
   return (
-    <Paper elevation={3}>
-      <img src={cover} alt="Book Cover" style={{ width: '100%' }} />
+    <Paper elevation={3} sx={{ padding: 0, margin: 0 }}>
+      <img src={cover} alt="Book Cover" style={{ width: '100%', height: '100%', display: 'block' }} />
     </Paper>
   );
 };
